Move to commands package and fix some compile errors
