+++ src/main/java/org/spongepowered/api/Game.java
@@ -27,8 +27,8 @@ package org.spongepowered.api;
 import org.spongepowered.api.asset.AssetManager;
 import org.spongepowered.api.command.CommandManager;
 import org.spongepowered.api.config.ConfigManager;
-import org.spongepowered.api.text.conversation.Conversation;
-import org.spongepowered.api.text.conversation.ConversationManager;
+import org.spongepowered.api.command.conversation.Conversation;
+import org.spongepowered.api.command.conversation.ConversationManager;
 import org.spongepowered.api.data.DataManager;
 import org.spongepowered.api.data.DataSerializable;
 import org.spongepowered.api.data.persistence.DataBuilder;
