+++ src/main/java/org/spongepowered/api/command/args/ResponseContext.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.text.conversation;
+package org.spongepowered.api.command.args;
 
 import org.spongepowered.api.text.Text;
