+++ src/main/java/org/spongepowered/api/command/conversation/Conversation.java
@@ -22,9 +22,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.text.conversation;
+package org.spongepowered.api.command.conversation;
 
-import org.spongepowered.api.text.conversation.QuestionResult.QuestionResultType;
+import org.spongepowered.api.command.conversation.QuestionResult.QuestionResultType;
 import org.spongepowered.api.data.DataContainer;
 import org.spongepowered.api.event.cause.Cause;
 import org.spongepowered.api.event.cause.conversation.ConversationEndType;
