+++ src/main/java/org/spongepowered/api/command/conversation/ConversationManager.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.text.conversation;
+package org.spongepowered.api.command.conversation;
 
 import org.spongepowered.api.event.cause.Cause;
 import org.spongepowered.api.event.cause.conversation.ConversationEndType;
